[CHORE] skybox 테스트중

+++ Source/GA6thFinal_Framework/Shaders/cs_cube_texture_convertor.hlsl
@@ -6,7 +6,7 @@ cbuffer CubeMapInfo
 {
     uint cubeFaceIndex;
     uint resolution;
-    float2 padding;
+    uint2 padding;
 };
 
 [numthreads(8, 8, 1)]
